+++ src/main/java/ai/lucidtech/las/sdk/ModelStatus.java
@@ -1,0 +1,14 @@
+package ai.lucidtech.las.sdk;
+
+
+public enum ModelStatus {
+    INACTIVE("inactive"),
+    ACTIVE("active"),
+    TRAINING("training");
+
+    public final String value;
+
+    ModelStatus(String value) {
+        this.value = value;
+    }
+}