+++ src/main/java/ai/lucidtech/las/sdk/NullableString.java
@@ -1,0 +1,17 @@
+package ai.lucidtech.las.sdk;
+
+
+class NullableString {
+    public Boolean hasEditedValue;
+    private String value;
+
+    public NullableString() {
+        this.hasEditedValue = false;
+        this.value = null;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+        this.hasEditedValue = true;
+    }
+}