+++ src/test/java/lucidtech/test/AppClientsTest.java
@@ -42,8 +42,19 @@ public class AppClientsTest extends ClientTest {
             .setDescription("AppClient Description")
             .setGenerateSecret(false)
             .setCallbackUrls(new String[] {"https://example.org/authCallback"})
-            .setLogoutUrls(new String[] {"https://example.org/logout"});
+            .setLogoutUrls(new String[] {"https://example.org/logout"})
+            .setLoginUrls(new String[] {"https://example.org/login"})
+            .setDefaultLoginUrl("https://example.org/login");
         JSONObject appClient = this.client.createAppClient(options);
+        this.assertAppClient(appClient);
+    }
+
+    @Test
+    public void testUpdateAppClient() throws IOException, APIException, MissingAccessTokenException {
+        UpdateAppClientOptions options = new UpdateAppClientOptions()
+            .setName("AppClient Name")
+            .setDescription("AppClient Description");
+        JSONObject appClient = this.client.updateAppClient(TestUtils.appClientId(), options);
         this.assertAppClient(appClient);
     }